+++ CHANGE_LOG.md
@@ -4,6 +4,7 @@
 - Support On-Premise licensing for OSATs, factories and secured facilities that are not connected to public internet
 - Upgrade the licensing
 - Removed logging specific options in favor of ILogger<T>
+- Merge `CommandString` into `CommandRoute`
 
 # 5.2.1-rc*
 > Breaking Change